argos/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
jinja2==3.1.3
aiofiles==23.2.1
python-docx==1.1.0
openai==1.12.0
python-dotenv==1.0.0
google-generativeai==0.8.0
anthropic==0.39.0
